JsonBadPlanTests: move to being regular tests
Makes the sanityCheckWorkers code public.
